added enum for user types
