+++ src/main/java/Classes/DAO/DatabaseSettings.java
@@ -1,0 +1,7 @@
+package Classes.DAO;
+
+public class DatabaseSettings {
+    public final static String DB_PATH = "/Users/stefanoverrilli/IdeaProjects/Pezzotammazon/db.sqlite";
+    public final static String DB_DRIVER = "org.sqlite.JDBC";
+    public final static String DB_TYPE = "sqlite";
+}
