--- src/main/java/Classes/ServletRegulation/AdminState.java
@@ -1,12 +1,0 @@
-package Classes.ServletRegulation;
-
-public class AdminState extends State{
-    public AdminState(Context context) {
-        super(context);
-    }
-
-    @Override
-    public void LoadLink() {
-        context.LoadAdminLinks();
-    }
-}
