--- src/main/java/Classes/ServletRegulation/State.java
@@ -1,11 +1,0 @@
-package Classes.ServletRegulation;
-
-public abstract class State {
-    protected Context context;
-
-    public State(Context context){
-        this.context = context;
-    }
-
-    public abstract void LoadLink();
-}
