--- src/main/java/Classes/ServletRegulation/UserState.java
@@ -1,12 +1,0 @@
-package Classes.ServletRegulation;
-
-public class UserState extends State{
-    public UserState(Context context) {
-        super(context);
-    }
-
-    @Override
-    public void LoadLink() {
-        context.LoadUserLinks();
-    }
-}
