+++ src/main/java/Classes/ServletsRegulation/AdminState.java
@@ -1,0 +1,12 @@
+package Classes.ServletsRegulation;
+
+public class AdminState extends State{
+    public AdminState(Context context) {
+        super(context);
+    }
+
+    @Override
+    public void LoadLink() {
+        context.LoadAdminLinks();
+    }
+}
