+++ src/main/java/Classes/ServletsRegulation/DefaultState.java
@@ -1,0 +1,15 @@
+package Classes.ServletsRegulation;
+
+import Classes.FrontController.Actions;
+
+public class DefaultState extends State{
+    public DefaultState(Context context) {
+        super(context);
+    }
+
+    @Override
+    public void LoadLink() {
+        Actions.delete();
+        context.LoadDefaultLinks();
+    }
+}
