+++ src/main/java/Classes/ServletsRegulation/State.java
@@ -1,0 +1,11 @@
+package Classes.ServletsRegulation;
+
+public abstract class State {
+    protected Context context;
+
+    public State(Context context){
+        this.context = context;
+    }
+
+    public abstract void LoadLink();
+}
