+++ src/main/java/Classes/ServletsRegulation/UserState.java
@@ -1,0 +1,12 @@
+package Classes.ServletsRegulation;
+
+public class UserState extends State{
+    public UserState(Context context) {
+        super(context);
+    }
+
+    @Override
+    public void LoadLink() {
+        context.LoadUserLinks();
+    }
+}
