+++ src/main/java/Classes/User/AccessLevels.java
@@ -1,0 +1,6 @@
+package Classes.User;
+
+public enum AccessLevels {
+    User,
+    Admin
+}
