--- src/main/java/Classes/User/ChainOfResponsability/Handler.java
@@ -1,11 +1,0 @@
-package Classes.User.ChainOfResponsability;
-
-import javax.servlet.http.HttpServletRequest;
-
-public abstract class Handler {
-    protected Handler next;
-    public void SetNext(Handler h){
-        this.next = h;
-    };
-    public abstract String handle(HttpServletRequest request) throws Exception;
-}
