--- src/main/java/Classes/User/ChainOfResponsability/RedirectHandler.java
@@ -1,19 +1,0 @@
-package Classes.User.ChainOfResponsability;
-
-import Classes.Command.DiscriminatorLinks;
-import Classes.Command.Dispatcher;
-import Classes.Command.Invoker;
-import Classes.User.UserModel;
-
-import javax.servlet.http.HttpServletRequest;
-
-public class RedirectHandler extends Handler{
-
-    @Override
-    public String handle(HttpServletRequest request) throws Exception {
-        DiscriminatorLinks discriminatorLinks = new DiscriminatorLinks();
-        UserModel user = (UserModel) request.getSession().getAttribute("user");
-        Invoker invoker = new Invoker(new Dispatcher(discriminatorLinks, user));
-        return invoker.executeOperation();
-    }
-}
