+++ src/main/java/Classes/User/ChainofChecks/NavbarCreationHandler.java
@@ -1,0 +1,13 @@
+package Classes.User.ChainofChecks;
+
+import Classes.Navbar.BuildNavbar;
+
+import javax.servlet.http.HttpServletRequest;
+
+public class NavbarCreationHandler extends Handler{
+    @Override
+    public String handle(HttpServletRequest request) throws Exception {
+        BuildNavbar.GetNavbar(request);
+        return next.handle(request);
+    }
+}
